resume-analyzer/backend/requirements.txt
fastapi
uvicorn[standard]
motor
spacy
sentence-transformers
openai
pymupdf
python-docx
python-dotenv
python-multipart
pydantic<2.0.0
setuptools
wheel
